# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "SellerInfo": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "categoryName": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "businessEmail": {
                        "type": "string"
                    },
                    "categoryId": {
                        "type": "string"
                    },
                    "itemCategories": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "logoUrl": {
                        "type": "string"
                    },
                    "websiteUrl": {
                        "type": "string"
                    },
                    "itemsCount": {
                        "type": "number"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "categoryName",
                    "name",
                    "description",
                    "phoneNumber",
                    "address",
                    "businessEmail",
                    "categoryId",
                    "itemCategories",
                    "logoUrl",
                    "websiteUrl",
                    "itemsCount"
                ]
            },
...
            "CategoryUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 5
                    },
                    "type": {
                        "enum": [
                            "Item",
                            "Service"
                        ],
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 250,
                        "minLength": 10
                    }
                },
                "type": "object"
            },
...
            "ProfileUpdateRequest": {
                "properties": {
                    "firstName": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 2
                    },
                    "lastName": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 2
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 150,
                        "minLength": 15
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 2
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 50
                    },
                    "businessEmail": {
                        "format": "email",
                        "type": "string"
                    },
                    "logoUrl": {
                        "format": "url",
                        "type": "string"
                    },
                    "websiteUrl": {
                        "format": "url",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "title": "sustainable-community-server API",
        "version": "1.0.0",
        "description": "Generated with `routing-controllers-openapi`"
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/sellers/profile": {
            "put": {
                "operationId": "SellerController.updateProfile",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileUpdateRequest"
                            }
                        }
                    },
                    "description": "ProfileUpdateRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Update seller profile",
                "tags": [
                    "Seller"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/sellers/{id}/categories/{categoryId}": {
            "put": {
                "operationId": "SellerController.updateCategory",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    },
                    {
                        "in": "path",
                        "name": "categoryId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Update seller category",
                "tags": [
                    "Seller"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/sellers/{id}/item-categories/{categoryId}": {
            "post": {
                "operationId": "SellerController.assignItemCategory",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    },
                    {
                        "in": "path",
                        "name": "categoryId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Assign item category to seller",
                "tags": [
                    "Seller"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "operationId": "SellerController.removeItemCategory",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    },
                    {
                        "in": "path",
                        "name": "categoryId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Remove item category from seller",
                "tags": [
                    "Seller"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/sellers/{id}/item-categories": {
            "get": {
                "operationId": "SellerController.getItemCategories",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Get seller's item categories",
                "tags": [
                    "Seller"
                ]
            }
        },
        "/sellers/": {
            "get": {
                "operationId": "SellerController.getListOfSellers",
                "parameters": [
                    {
                        "in": "query",
                        "name": "activeSellers",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/SellerInfo"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Get list of sellers",
                "tags": [
                    "Seller"
                ]
            }
        },
        "/sellers/{id}": {
            "get": {
                "operationId": "SellerController.getSeller",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SellerInfo"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Get seller by id",
                "tags": [
                    "Seller"
                ]
            }
        },
...
        "/customers/profile": {
            "put": {
                "operationId": "CustomerController.updateProfile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileUpdateRequest"
                            }
                        }
                    },
                    "description": "ProfileUpdateRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Update profile",
                "tags": [
                    "Customer"
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "operationId": "CategoryController.getOneCategory",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Get category by id",
                "tags": [
                    "Category"
                ]
            },
            "put": {
                "operationId": "CategoryController.updateCategory",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryUpdateRequest"
                            }
                        }
                    },
                    "description": "CategoryUpdateRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Update a category",
                "tags": [
                    "Category"
                ]
            },
            "delete": {
                "operationId": "CategoryController.deleteCategory",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Delete a category",
                "tags": [
                    "Category"
                ]
            }
        },
...
    }
}